api/requirements.txt
flask~=1.1.1
flask-restful~=0.3.7
flask-sqlalchemy~=2.4.1
flask-migrate~=2.5.2
flask-cors~=3.0.8
flask-sse~=0.2.1
mysql-connector-python~=8.0.18
mysqlclient~=2.0.3
numpy~=1.18.0
sympy~=1.5
h5py~=2.10.0
SQLAlchemy~=1.3.12
python-dateutil~=2.8.1
celery~=4.4.7
cantools~=35.3.0
redis~=3.5.3
gunicorn~=20.0.4
gevent~=20.9.0
boto3~=1.15.18
python-dotenv~=0.14.0
pandas~=1.2.0
